refactor: move to `/api/assignments`
